refactor(mypage): 마이페이지 컴포넌트 폴더 구조 변경
